Fix Bug

+++ app/Console/Commands/CrudGenerator.php
@@ -111,28 +111,31 @@ class CrudGenerator extends Command
         $str = str_replace("{{ deleteAction }}", $this->deleteAction, $str);
         file_put_contents($controllerFile, $str);
 
-        //Create Store Action
-        $storeActionFile = app_path() . "/Actions/" . $name . '/Store.php';
-        $storeAction = file_get_contents($storeActionFile);
-        $storeAction = str_replace('//use .. ;', "use App\\Models\\" . $this->model . ';', $storeAction);
-        $storeAction = str_replace('public function handle($handle)', 'public function handle(Array $request)', $storeAction);
-        $storeAction = str_replace('// ..', '$handle = ' . $this->model . '::create($request);', $storeAction);
-        file_put_contents($storeActionFile, $storeAction);
+        if ($this->option('action')) {
+            //Create Store Action
+            $storeActionFile = app_path() . "/Actions/" . $name . '/Store.php';
+            $storeAction = file_get_contents($storeActionFile);
+            $storeAction = str_replace('//use .. ;', "use App\\Models\\" . $this->model . ';', $storeAction);
+            $storeAction = str_replace('public function handle($handle)', 'public function handle(Array $request)', $storeAction);
+            $storeAction = str_replace('// ..', '$handle = ' . $this->model . '::create($request);', $storeAction);
+            file_put_contents($storeActionFile, $storeAction);
 
-        //Create Update Action
-        $updateActionFile = app_path() . "/Actions/" . $name . '/Update.php';
-        $updateAction = file_get_contents($updateActionFile);
-        $updateAction = str_replace('//use .. ;', "use App\\Models\\" . $this->model . ';', $updateAction);
-        $updateAction = str_replace('public function handle($handle)', 'public function handle(' . $this->model . ' $' . $this->varModel . ', Array $request)', $updateAction);
-        $updateAction = str_replace('// ..', '$handle = $' . $this->varModel . '->update($request);', $updateAction);
-        file_put_contents($updateActionFile, $updateAction);
 
-        //Create delete Action
-        $deleteActionFile = app_path() . "/Actions/" . $name . '/Delete.php';
-        $deleteAction = file_get_contents($deleteActionFile);
-        $deleteAction = str_replace('//use .. ;', "use App\\Models\\" . $this->model . ';', $deleteAction);
-        $deleteAction = str_replace('public function handle($handle)', 'public function handle(' . $this->model . ' $' . $this->varModel . ')', $deleteAction);
-        $deleteAction = str_replace('// ..', '$handle = $' . $this->varModel . '->delete();', $deleteAction);
-        file_put_contents($deleteActionFile, $deleteAction);
+            //Create Update Action
+            $updateActionFile = app_path() . "/Actions/" . $name . '/Update.php';
+            $updateAction = file_get_contents($updateActionFile);
+            $updateAction = str_replace('//use .. ;', "use App\\Models\\" . $this->model . ';', $updateAction);
+            $updateAction = str_replace('public function handle($handle)', 'public function handle(' . $this->model . ' $' . $this->varModel . ', Array $request)', $updateAction);
+            $updateAction = str_replace('// ..', '$handle = $' . $this->varModel . '->update($request);', $updateAction);
+            file_put_contents($updateActionFile, $updateAction);
+
+            //Create delete Action
+            $deleteActionFile = app_path() . "/Actions/" . $name . '/Delete.php';
+            $deleteAction = file_get_contents($deleteActionFile);
+            $deleteAction = str_replace('//use .. ;', "use App\\Models\\" . $this->model . ';', $deleteAction);
+            $deleteAction = str_replace('public function handle($handle)', 'public function handle(' . $this->model . ' $' . $this->varModel . ')', $deleteAction);
+            $deleteAction = str_replace('// ..', '$handle = $' . $this->varModel . '->delete();', $deleteAction);
+            file_put_contents($deleteActionFile, $deleteAction);
+        }
     }
 }
